+++ Sources/RichTextKit/Commands/RichTextCommandsFontSizeOptionsGroup.swift
@@ -9,8 +9,7 @@
 import SwiftUI
 
 /**
- This view can be added to a `Commands`, to expose rich text
- font size menu options to an app main menu or key commands.
+ This view defines `Commands` content for font size options.
  */
 public struct RichTextCommandsFontSizeOptionsGroup: View {