+++ Sources/RichTextKit/Commands/RichTextCommandsIndentOptionsGroup.swift
@@ -9,8 +9,7 @@
 import SwiftUI
 
 /**
- This view can be added to a `Commands`, to expose rich text
- indent options to an app main menu or key commands.
+ This view defines `Commands` content for indent options.
  */
 public struct RichTextCommandsIndentOptionsGroup: View {